+++ sql/core/src/main/scala/org/apache/spark/sql/execution/window/BoundOrdering.scala
@@ -25,18 +25,22 @@ import org.apache.spark.sql.catalyst.expressions.Projection
  * Function for comparing boundary values.
  */
 private[window] abstract class BoundOrdering {
-  def compare(inputRow: InternalRow, inputIndex: Int, outputRow: InternalRow, outputIndex: Int): Int
+  def compare(
+      inputRow: InternalRow,
+      inputIndex: Long,
+      outputRow: InternalRow,
+      outputIndex: Long): Long
 }
 
 /**
  * Compare the input index to the bound of the output index.
  */
-private[window] final case class RowBoundOrdering(offset: Int) extends BoundOrdering {
+private[window] final case class RowBoundOrdering(offset: Long) extends BoundOrdering {
   override def compare(
       inputRow: InternalRow,
-      inputIndex: Int,
+      inputIndex: Long,
       outputRow: InternalRow,
-      outputIndex: Int): Int =
+      outputIndex: Long): Long =
     inputIndex - (outputIndex + offset)
 }
 
@@ -51,8 +55,8 @@ private[window] final case class RangeBoundOrdering(
 
   override def compare(
       inputRow: InternalRow,
-      inputIndex: Int,
+      inputIndex: Long,
       outputRow: InternalRow,
-      outputIndex: Int): Int =
+      outputIndex: Long): Long =
     ordering.compare(current(inputRow), bound(outputRow))
 }
